Add wpt test for font feature precedence for 'dlig'.
The new Lato-Medium-Liga font is based on Lato-Medium 3.002 available
at https://github.com/latofonts/lato-source/tree/master/fonts/static/TTF
It is subset with hb-subset with "fistﬆﬁ " and the 'fi' ligatures
moved back to 'liga' (version 3.002 moved them to 'dlig'). The crossbar
of 't' in the 'st' ligature was also shorted to ensure that it would not
cover a normal 't'.

The Lato-Medium currently available to the tests is version 2.015 and
lacks a mapping from U+fb06 to the 'st' ligature glyph which was only
accessible through the 'dlig' feature. Lato-Medium 3.002 added this
mapping (required for this new test) but also moved the 'fi' ligatures
to 'dlig' and some existing tests depend on the 'fi' ligatures being in
'liga'.

This tests https://github.com/w3c/csswg-drafts/pull/3430

Change-Id: Iaf0ff1f99f3ec282bacb0e3cb44ede27d47fac5a