[COOP] Access reporting. Fix the python server.
A python server is used has a communication channel in between documents
from different browsing context group and potentially cross-origin.

There are two methods:
- one for writing a message.
- one for reading the message.

I initially supposed we won't need to store more than one message in the
stash. It turns out there are test doing two write in row without any
read in the middle. Moreover, this was throwing an exception when this
happened.

This patch makes the server to maintain a queue of messages instead and
stop throwing an error.

Bug:chromium:1109194
Change-Id: I25f8dc2c2a032781a4527a0994e94e46f85e12c9