Trust Tokens: Alter signing interface to take a list of issuers
This is the first of two changes altering Trust Tokens to take a list of
issuers, rather than a single issuer, as an argument to its signing
operation. This first change updates the interface and code passing the
parameter's value from Blink to the network service, but keeps the
current network service functionality the same (choosing the first
issuer provided in the list, if any, and not updating the format of the
Sec-Signature / Sec-Signed-Redemption-Records constructed during the
signing operation). The second change will update the code constructing
the outgoing headers to accommodate multiple issuers.

This is a backwards-incompatible change to the interface, but Trust
Tokens is currently experimental and provides no version-to-version
compatibility guarantees.

More at
https://docs.google.com/document/u/1/d/1L7OPuLt9CE5VhT_bGGwp3xej3mehGF-mS1ZZOVX8yp4/edit#

Test: Updates iframe-attribute parsing tests, WPTs, and browser tests
Bug: 1105230
Change-Id: I00a81758a0412f3e93ad150dc4736e55c9d9bbd3