Fix onbeforexrselect event handler inheritance
The onbeforexrselect event is supposed to be part of GlobalEventHandlers
and separately inherited by {HTML,MathML,SVG}Element, Window, and
Document. The current implementation had added it at the Element
level which isn't correct, causing test failures in the idlharness
test.

Update the implementation to fix the inheritance, which unfortunately
adds a bit of extra boilerplate.

Also add objects to the idlharness test to verify the inheritance.

Spec: https://immersive-web.github.io/dom-overlays/#onbeforexrselect

Fixed: 1108685
Change-Id: I8df73304b7b2d9bd93df30b78953aa9b378f8cad